Add reference to default CSP directives
Helmet has a contentSecurityPolicy property that can be used to get the
default directives for the CSP middleware. A reference to this object
has been added here to make the default Content Security Policy
directives available in fastify-helmet and to make sure the examples in
helmet's documentation also work for fastify-helmet.